fix(KX1): envelope times were 1000x too fast
The rate table computed

    seconds = 0.002f * powf(3000.0f, i / 15.0f) / 1000.0f

but 0.002 is already in seconds, so the trailing /1000 shrank the entire
curve by three orders of magnitude. The documented "nibble 0 = ~2ms,
nibble 15 = ~6s" range actually ran 2us .. 6ms: every note was a click,
and nibbles 0-5 all saturated at one sample per step, so most of the
ADSR range was not merely wrong but indistinguishable.

Removing the /1000 is necessary but not sufficient. At 44.1 kHz a 6
second sweep needs a per-sample step of 65535/264600 = 0.25, which
truncates to 0 in the uint16_t accumulator; clamping it up to 1 — as the
old code did — pins every envelope slower than ~1.5 s to that same
~1.5 s. So the envelope domain gains 8 fractional bits (Q8), which holds
the full span to within 0.6%. kEnvMax*2 still fits in int32, so the
audio path stays 32-bit.

Measured attack/release, previously 0.02 ms .. 6.03 ms with the fast
half of the range collapsed onto a single sample:

    nibble  0     2.0 ms        nibble  8     143.1 ms
    nibble  4    16.9 ms        nibble 12    1211.6 ms
    nibble  7    83.9 ms        nibble 15    6038.6 ms

Levels are stored Q8 and narrowed to the 0..0xFFFF amplitude before the
mixing multiplies, which keeps those products in their original range —
multiplying the widened level directly would overflow int32.

Tests: S5/S5b check both endpoints and require every nibble to be at
least 25% slower than its predecessor, which is what the old table
failed. Verified to fail both when the /1000 is reinstated and when the
Q8 precision alone is removed.

+++ sources/Application/Instruments/SynthVoice.h
@@ -136,15 +136,18 @@ struct SynthVoice {
   uint32_t noise_state; // 23-bit LFSR
 
   // --- operator envelopes ---
-  uint16_t op1_env_level, op2_env_level, op3_env_level;
+  // Levels are Q8: the low 8 bits are fractional headroom, without which a
+  // multi-second envelope's per-sample step truncates to zero. Consumers
+  // shift down by kEnvFracBits to get the 0..0xFFFF amplitude.
+  uint32_t op1_env_level, op2_env_level, op3_env_level;
   uint8_t op1_env_stage, op2_env_stage, op3_env_stage;
 
   // --- filter envelope ---
-  uint16_t filt_env_level;
+  uint32_t filt_env_level;
   uint8_t filt_env_stage;
 
   // --- pitch envelope (AD only) ---
-  uint16_t pitch_env_level;
+  uint32_t pitch_env_level;
   uint8_t pitch_env_stage;
 
   // --- state variable filter ---
